Move mock generation into package_test.go, and move TestPackage into upgradedatabase.
